new feature:
File mappings can be done. A to be mapped file will be replaced while a fulltext search by the file in fieMappings.xml in the result lists of articles.
This can be used to do a fulltext search in OCR-images that have a belonging text file, which contains the text.

E.g.: hit = "bild.txt" => mapping from "txt" to "jpg" : hit = "bild.jpg"